+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- When using a volume source (snapshot or other volume), LINSTOR CSI now correctly resizes the volume to the new
+  requested size.
+
 ## [0.16.0] - 2021-10-15
 
 ### Changed